requirements.txt
azure-functions
deltalake
pandas